requirements.txt
flask
pycryptodome
python-dotenv
